requirements.txt
# Core
fastapi==0.109.0
uvicorn==0.27.0
gunicorn==21.2.0

# Multipart form parsing for UploadFile
python-multipart>=0.0.5

# Database
sqlalchemy==2.0.25
pyodbc==5.1.0
python-dotenv==1.0.0
azure-cosmos>=4.5.0

# AI / HTTP
requests==2.31.0
httpx>=0.24.0

# Config
pydantic-settings==2.0.3

# Data processing
pandas>=2.0.0
numpy>=1.24.0
openpyxl>=3.0.0
xlrd>=2.0.1
azure-storage-blob>=12.0.0
sentence-transformers>=2.2.2
